Refactor to move registration and js to internal packages
