+++ plugins/plotly-express/setup.cfg
@@ -34,4 +34,4 @@ where=src
 
 [options.entry_points]
 deephaven.plugin =
-    registration_cls = deephaven.plot.express:ExpressRegistration
+    registration_cls = deephaven.plot.express._register:ExpressRegistration
