+++ plugins/plotly-express/src/deephaven/plot/express/_register.py
@@ -1,0 +1,24 @@
+from . import DeephavenFigureType
+from ._js import create_js_plugin
+
+from deephaven.plugin import Registration, Callback
+
+
+class ExpressRegistration(Registration):
+    """
+    Register the DeephavenFigureType and a JsPlugin
+
+    """
+
+    @classmethod
+    def register_into(cls, callback: Callback) -> None:
+        """
+        Register the DeephavenFigureType and a JsPlugin
+
+        Args:
+          callback: Registration.Callback:
+            A function to call after registration
+
+        """
+        callback.register(DeephavenFigureType)
+        callback.register(create_js_plugin())
